Map of collection sites (USA)
I wanted a way to visualize where these sites are and thought this would be useful to everyone. 

This is a map of the rough locations of the traps. It is based off of the list @ "https://github.com/ReproducibleQM/sapsuckRz/blob/master/counties%2Bcoordinates2.csv". I either searched the coordinates or the city/state. For the KBS site I put it in hickory corners. The linked image doesn't show all the sites, but the ones that are not highlighted are within the range of what is outlined here. You can use the link at the bottom to see where the other sites are.

I used a handy website to create this map, "MapFling" and the following website will recapitulate the original map that I snapped this image from http://mapfling.com/qsmw5zd